Add files via upload
Photopigment Database Updates
August 14, 2024

Per communication with Joshua Grosse.

Anolis carolinensis is the main one. You quote the values 560,495,437,358 from Kawamura & Yokoyama 1998. However, in that paper it mentions these are the λmax values for the opsins recombined with A1. In the actual animal they are combined with A2, and so given the notably longer wavelengths 625,503,462,365.  

– Updated

Xenopus laevis is listed with rods at 524,445,434. The first two values are using A2 but the last is for the SWS1 pigment recombined with A1, so it would have a somewhat lower wavelength too.

– Updated 434 to 443 nm in presence of A2.  Corrected 445 nm to 425 nm, which is a cone, not a rod.  See summary in https://www.frontiersin.org/journals/neuroanatomy/articles/10.3389/fnana.2021.784478/full 

Gekko gecko is listed with the values as for cones, but this is a nocturnal species and has an all-rod retina.  

 – Updated to rods

I think the same is true for Hemidactylus species.  

– Confirmed nocturnal and also updated to rods.

You attribute the values 520,466,365 for H. frenatus to Loew et al. 1996 but it doesn't seem like they all come from there.  

--  Terascincus scincus was in Loew et al. 1996 and omitted from the database.  It is now added.  H. frenatus is found in Crescitelli et al. 1977 J. Physiol. 268, 559–573, except for the UV pigment. 


Coccinella septempunctata is given two values, 570 and 370 nm following Lin & Wu 1992. That paper also indicated that they had a third sensitivity though, and the next year Lin has the paper "Identification of photoreceptor locations in the compound eye of Coccinella septempunctata" that gives values 520, 420, 360 nm. 

–	Updated. 

Taxonomic Updates 

•  Anthophora is listed as Adrenidae instead of Apidae
•  Colletes is listed as Apidae (and as Collectes), instead of Colletidae
•  Formica is listed as Colletidae and Cataglyphis as Crabronidae, instead of Formicidae
•  Anthidium is listed as Formicidae, instead of Megachilidae
•  Cerceris is listed as Sphecidae but seems like it belongs in Philanthidae / Philanthinae in Crabronidae, wherever you have Philanthus


These have been updated / corrected.